captcha/requirements.txt
fastapi
numpy
opencv-python
pillow
pytesseract
requests
pytest
requests-mock
pytz